+++ lib/Controller/SettingsController.php
@@ -14,12 +14,28 @@ class SettingsController extends Controller
     private $config;
     private $userId;
 
+    protected $isAdmin;
+
     public function __construct(IRequest $request, IConfig $config, string $userId)
     {
         parent::__construct(Application::APP_ID, $request);
 
         $this->config = $config;
         $this->userId = $userId;
+        $this->isAdmin = $this->request->getParam('type') === 'admin';
+    }
+
+    /**
+     * @NoAdminRequired
+     */
+    public function index()
+    {
+        return new JSONResponse([
+            Application::SETTINGS_KEY_DEFAULT_ENABLE      => $this->config->getAppValue(Application::APP_ID, Application::SETTINGS_KEY_DEFAULT_ENABLE),
+            Application::SETTINGS_KEY_ENABLE              => $this->config->getUserValue($this->userId, Application::APP_ID, Application::SETTINGS_KEY_ENABLE),
+            Application::SETTINGS_KEY_DEFAULT_COPY_PREFIX => $this->config->getAppValue(Application::APP_ID, Application::SETTINGS_KEY_DEFAULT_COPY_PREFIX),
+            Application::SETTINGS_KEY_COPY_PREFIX         => $this->config->getUserValue($this->userId, Application::APP_ID, Application::SETTINGS_KEY_COPY_PREFIX),
+        ]);
     }
 
     /**
@@ -27,7 +43,40 @@ class SettingsController extends Controller
      */
     public function enable(string $enabled)
     {
-        $this->config->setUserValue($this->userId, Application::APP_ID, Application::SETTINGS_KEY_ENABLE, $enabled);
+        if ($this->isAdmin) {
+            $this->config->setAppValue(Application::APP_ID, Application::SETTINGS_KEY_DEFAULT_ENABLE, $enabled);
+        } else {
+            $this->config->setUserValue($this->userId, Application::APP_ID, Application::SETTINGS_KEY_ENABLE, $enabled);
+        }
+
+        return new JSONResponse();
+    }
+
+    /**
+     * @NoAdminRequired
+     */
+    public function setCopyPrefix(string $prefix)
+    {
+        if ($this->isAdmin) {
+            $this->config->setAppValue(Application::APP_ID, Application::SETTINGS_KEY_DEFAULT_COPY_PREFIX, trim($prefix));
+        } else {
+            $this->config->setUserValue($this->userId, Application::APP_ID, Application::SETTINGS_KEY_COPY_PREFIX, trim($prefix));
+        }
+
+        return new JSONResponse();
+    }
+
+    /**
+     * @NoAdminRequired
+     */
+    public function setSharingFolder(string $folder)
+    {
+        if ($this->isAdmin) {
+            $this->config->setAppValue(Application::APP_ID, Application::SETTINGS_KEY_DEFAULT_SHARING_FOLDER, trim($folder));
+        } else {
+            $this->config->setUserValue($this->userId, Application::APP_ID, Application::SETTINGS_KEY_SHARING_FOLDER, trim($folder));
+        }
+
         return new JSONResponse();
     }
 }